First stupidly simple version
This version is very static: player names can't be changed, it is not
possible to alter scores once they are entered, and the game moves
forward in steps by clicking on buttons. It also hides the bids and
results (except for the actual scores) once a round has finished. CSS,
or layout in general, are also severely lacking.